requirements.txt
selenium==4.25.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
schedule==1.2.2